Extract test data classes
