+++ test/ForEvolve.EntityFrameworkCore.Tests/Seeders/SeederManagerTest.cs
@@ -125,84 +125,4 @@ namespace ForEvolve.EntityFrameworkCore.Seeders
             }
         }
     }
-
-    namespace TestData
-    {
-        public class SeederDbContext : DbContext
-        {
-            private SeederDbContext(DbContextOptions options) : base(options)
-            {
-                _database = new TestDatabaseFacade(this);
-            }
-
-            public DbSet<TestEntity> TestEntities { get; set; }
-
-            public static SeederDbContext Create(string databaseName)
-            {
-                var builder = new DbContextOptionsBuilder<SeederDbContext>()
-                    .UseInMemoryDatabase(databaseName);
-                return new SeederDbContext(builder.Options);
-            }
-
-            public int BeginTransactionCalls => _database.BeginTransactionCalls;
-            public int CommitTransactionCalls => _database.CommitTransactionCalls;
-            public int RollbackTransactionCalls => _database.RollbackTransactionCalls;
-            public int SaveChangesCalls { get; private set; }
-
-
-            private readonly TestDatabaseFacade _database;
-            public override DatabaseFacade Database => _database;
-
-            public bool SaveChangesShouldThrow => SaveChangesException != default;
-            public Exception SaveChangesException { get; set; }
-
-            public override int SaveChanges()
-            {
-                if (SaveChangesShouldThrow)
-                {
-                    throw SaveChangesException;
-                }
-                SaveChangesCalls++;
-                return 1;
-            }
-        }
-
-        public class TestDatabaseFacade : DatabaseFacade
-        {
-            public TestDatabaseFacade(DbContext context) : base(context) { }
-
-            public int BeginTransactionCalls { get; private set; }
-            public int CommitTransactionCalls { get; private set; }
-            public int RollbackTransactionCalls { get; private set; }
-
-            public override IDbContextTransaction BeginTransaction()
-            {
-                var transactionMock = new Mock<IDbContextTransaction>();
-                transactionMock
-                    .Setup(s => s.Commit())
-                    .Callback(() => CommitTransactionCalls++);
-                transactionMock
-                    .Setup(s => s.Rollback())
-                    .Callback(() => RollbackTransactionCalls++);
-                BeginTransactionCalls++;
-                return transactionMock.Object;
-            }
-
-            public override void CommitTransaction()
-            {
-                CommitTransactionCalls++;
-            }
-
-            public override void RollbackTransaction()
-            {
-                RollbackTransactionCalls++;
-            }
-        }
-
-        public class TestEntity
-        {
-            public int Id { get; set; }
-            public string Name { get; set; }
-        }
-    }
 }
