+++ test/ForEvolve.EntityFrameworkCore.Tests/Seeders/TestData/TestEntity.cs
@@ -1,0 +1,8 @@
+﻿namespace ForEvolve.EntityFrameworkCore.Seeders.TestData
+{
+    public class TestEntity
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}
